requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.0.0
apturl==0.5.2
argcomplete==1.8.1
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
atomicwrites==1.1.5
attrs==19.3.0
audioread==3.0.1
autobahn==17.10.1
Automat==0.8.0
backoff==2.2.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
bloom==0.11.2
breezy==3.0.2
Brlapi==0.7.0
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
cbor==1.0.0
certifi==2019.11.28
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.0
chromadb==0.3.29
ci-sdr==0.0.2
Click==7.0
clickhouse-connect==0.6.14
cloud-init==23.3.1
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.1.1
colcon-cmake==0.2.28
colcon-common-extensions==0.3.0
colcon-core==0.15.0
colcon-defaults==0.2.8
colcon-devtools==0.2.4
colcon-installed-package-information==0.1.0
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.2.15
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.3.3
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.8
colcon-recursive-crawl==0.2.3
colcon-ros==0.4.1
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.3
coloredlogs==15.0.1
command-not-found==0.3
ConfigArgParse==1.7
configobj==5.0.6
constantly==15.1.0
cov-core==1.15.0
coverage==4.5.2
cryptography==2.8
ctc-segmentation==1.7.4
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.14
dataclasses-json==0.6.1
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
Deprecated==1.2.7
Distance==0.1.3
distlib==0.3.0
distro==1.4.0
distro-info==0.23+ubuntu1.1
docutils==0.16
duckdb==0.9.1
dulwich==0.19.15
duplicity==0.8.12.0
editdistance==0.6.2
einops==0.7.0
emoji==2.8.0
empy==3.3.2
entrypoints==0.3
espnet==202308
espnet-model-zoo==0.1.7
espnet-tts-frontend==0.0.3
exceptiongroup==1.1.3
fast-bss-eval==0.1.3
fastapi==0.85.1
fasteners==0.14.1
fastimport==0.9.8
ffmpeg-python==0.2.0
filelock==3.12.4
filetype==1.2.0
flake8==3.7.9
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.2
future==0.18.2
g2p-en==2.1.0
gdown==4.4.0
gpg==1.13.1
graphlib-backport==1.0.3
greenlet==3.0.0
h11==0.14.0
h5py==3.10.0
hnswlib==0.7.0
html5lib==1.0.1
httplib2==0.14.0
httptools==0.6.0
huggingface-hub==0.18.0
humanfriendly==10.0
hydra-core==1.3.2
hyperlink==19.0.0
idna==2.8
ifcfg==0.18
importlib-metadata==1.5.0
importlib-resources==6.1.0
incremental==16.10.1
inflect==7.0.0
inflection==0.5.1
jaconv==0.3.4
jamo==0.4.1
Jinja2==2.10.1
joblib==1.3.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
kaldiio==2.18.0
keyring==18.0.1
kiwisolver==1.0.1
langchain==0.0.291
langdetect==1.0.9
langsmith==0.0.43
language-selector==0.1
lark-parser==0.8.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
librosa==0.9.2
llama-index==0.8.32
llvmlite==0.41.0
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
markdown-it-py==3.0.0
MarkupSafe==1.1.0
marshmallow==3.20.1
matplotlib==3.1.2
mccabe==0.6.1
mdurl==0.1.2
mock==3.0.5
monotonic==1.5
more-itertools==4.2.0
mpi4py==3.0.3
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
netifaces==0.10.4
networkx==3.1
nltk==3.8.1
nose2==0.9.1
notify2==0.3
numba==0.58.0
numexpr==2.8.6
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
omegaconf==2.3.0
onnxruntime==1.16.1
openai==0.28.1
openai-whisper==20230314
opt-einsum==3.3.0
overrides==7.4.0
packaging==20.3
pandas==2.0.3
parallel-wavegan==0.5.4
paramiko==2.6.0
pbr==5.4.5
pexpect==4.6.0
Pillow==7.0.0
platformdirs==3.11.0
pluggy==0.13.0
pooch==1.7.0
posthog==3.0.2
protobuf==3.6.1
psutil==5.5.1
pulsar-client==3.3.0
py==1.8.1
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.13
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.21
pycups==1.9.73
pydantic==2.4.2
pydantic-core==2.10.1
pydocstyle==2.1.1
pydot==1.4.1
pydub==0.25.1
pyflakes==2.1.1
PyGithub==1.43.7
Pygments==2.16.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.1.0
PyNaCl==1.3.0
pyopenjtalk==0.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypinyin==0.44.0
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
pytest==4.6.9
pytest-cov==2.8.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.0
python-gitlab==2.0.1
python-iso639==2023.6.15
python-magic==0.4.27
python-snappy==0.5.3
pytorch-wpe==0.0.1
PyTrie==0.2
pytz==2023.3.post1
pyusb==1.2.1
pyworld==0.3.4
pyxdg==0.26
PyYAML==5.3.1
regex==2023.10.3
reportlab==3.5.34
requests==2.31.0
requests-unixsocket==0.2.0
resampy==0.4.2
rich==13.5.3
robotics-audio==0.1.0
roman==2.0.0
rosdep==0.22.2
rosdep-modules==0.22.2
rosdistro==0.9.0
rosdistro-modules==0.9.0
rospkg-modules==1.5.0
scikit-learn==1.3.1
scipy==1.10.1
SecretStorage==2.3.1
sentencepiece==0.1.97
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sniffio==1.3.0
snowballstemmer==2.0.0
sos==4.5.6
soundfile==0.12.1
soupsieve==1.9.5
SpeechRecognition==3.10.0
SQLAlchemy==2.0.20
ssh-import-id==5.10
starlette==0.20.4
supervisor==4.2.5
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.2.3
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tiktoken==0.3.1
tokenizers==0.14.1
tomlkit==0.12.1
torch==2.1.0
torch-complex==0.4.3
tqdm==4.66.1
Twisted==18.9.0
txaio==2.10.0
typeguard==2.13.3
typing-extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
u-msgpack-python==2.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.3.7
unstructured==0.10.19
urllib3==1.25.8
uvicorn==0.23.2
uvloop==0.17.0
vcstool==0.3.0
vcstools==0.1.42
wadllib==1.3.3
watchfiles==0.20.0
wcwidth==0.1.8
webencodings==0.5.1
websockets==11.0.3
wrapt==1.11.2
wsaccel==0.6.2
xkit==0.0.0
xmltodict==0.13.0
yarl==1.9.2
yq==3.2.3
zipp==3.17.0
zope.interface==4.7.1
zstandard==0.21.0
